Initial venue details form refactor
