+++ src/components/atoms/AutoSpan/index.ts
@@ -1,0 +1,1 @@
+export { AutoSpan } from "components/atoms/AutoSpan/AutoSpan";
